Allow error GPUDevices to be returned
Removes asserts that expect a failed requestDevice to always return a
null device, and allow it to instead return the provided error device
which is immediately marked as lost.

Also updated the case where a stale adapter rejected with an exception
to instead force the device that was created to be lost and return it
instead. This is not ideal, as we'd rather have Dawn track the adapter's
stale state and return an error device, but in the meantime it gets
several new CTS tests passing.

Bug: chromium:1234617
Change-Id: I6220aca621f82922cc6317e2ec2ee6d9ca6ed762
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4550020
Reviewed-by: Austin Eng <enga@chromium.org>
Commit-Queue: Brandon Jones <bajones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1148853}
NOKEYCHECK=True
GitOrigin-RevId: 78877ace7aac989d9229b62019e9b81bb9ed4a65