Add WPT for calling window.fence methods in an inactive document.
This CL adds a test that ensures that window.fence methods properly
error out if the caller attempts to call it in a document that's not
active.

The test specifically checks the `setReportEventDataForAutomaticBeacons`
method. If that passes, we know that the other fence methods will also
not work, since they all rely on the same check.

Change-Id: I2eb6823e406a9751132ea7a3f776eafca499cd45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4551455
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Commit-Queue: Liam Brady <lbrady@google.com>
Cr-Commit-Position: refs/heads/main@{#1149155}
NOKEYCHECK=True
GitOrigin-RevId: 829795c9818146b3c13e6320ef7fdb220989a014