Remove incorrect/unnecessary forced style update
UpdateStyleAndLayoutTree() that was documented as updating for popover
transitions, but was done after the popover was added to the top layer
and marked as showing, which would not have such an effect.

Also, the html spec does not support this forced update. Removing.

Change-Id: I7e4410200c1e513c2acad23f926e05511848bd56
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4562593
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1149585}
NOKEYCHECK=True
GitOrigin-RevId: 2759da6c46b79866b630a60e8d9c266ba8ec43fd