bindings: Supports [DeprecateAs=] applied to IDL dictionary members
There has been one place (AuctionAdInterestGroup.dailyUpdateUrl)
that wanted to use [DeprecateAs=], but we've not supported it yet.
Plus, there is another interest to use [DeprecateAs=] on an IDL
dictionary member. Hence, this patch supports it.

The deprecation with [DeprecateAs=] will be counted when converting
a JS object to an IDL dictionary object and when converting an IDL
dictionary object to a JS object, rather than when accessing an IDL
dictionary member because a) accesses to IDL dictionary members
happen more often, and b) an ExecutionContext to store the count
is not available in accessors of IDL dictionary members.

Change-Id: I699bf8103394fca4e3eed3ffb059a525b255a01b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4567765
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Yuki Shiino <yukishiino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1149640}
NOKEYCHECK=True
GitOrigin-RevId: cfa869b6793d6a735e38adac2a55c645d6cae89f