COOP: restrict-properties reporting 3/*: Navigation reporting.
This patch is the first step in the implementation of reporting support
for COOP: restrict-properties. We now take into account COOP:
restrict-properties in the computation of virtual browsing context
group. To support swaps in the same coop related group for virtual
browsing context groups, we add the concept of a virtual coop related
group, and associate a virtual browsing context group with its virtual
coop related group.

This addresses the basic navigation reporting case. This patch does not
yet implement:
- Access reporting for COOP: restrict-properties.
- The virtual browsing context group reuse algorithm.

Low-Coverage-Reason: The introduced code in RenderFrameHostDelegate
is a simple empty return function.

Bug: 1424417
Change-Id: I195512d5e216175a8469d018dbdc9418f3e4575d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4523104
Commit-Queue: Arthur Hemery <ahemery@chromium.org>
Reviewed-by: Camille Lamy <clamy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1149665}
NOKEYCHECK=True
GitOrigin-RevId: 79d203e5d63487723ac7652e9b903b0b4395ae50