Change WebNavigationTimings.system_entropy_at_navigation_start default value to kEmpty
In a fuzzer test that creates a frame, we were seeing a check failure
because we ended up returning kNormal for a frame, when we should have
been returning kEmpty.

This change flips the default value for WebNavigationTimings to be
kEmpty instead of kNormal, which fixes the initial issue. However,
it exposes another condition where calling window.open('about:blank') returns in a kEmpty value. In these cases, we force the entropy type
to be kNormal.

Bug: 1446488

Change-Id: I3620786097c8581a1f3b97030d5d044e3f2c009e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4554162
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Commit-Queue: Mike Jackson <mjackson@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1149762}
NOKEYCHECK=True
GitOrigin-RevId: bb377506772eccede3b175c54e8a130a0f9a67f7