VT: Fix index ordering that can cause a potential tag order inversion.
This patch addresses the fact that we may have an element index overlap
between root and non-root elements, which is sorted with an unstable
sort.

We do this by adding 1000 to all of the non-root element indices when
they are in view_transition_style_tracker. We subtract out this offset
when adding the index to the view transition id, because cc needs to
match the elements based on an index in a vector.

R=bokan@chromium.org, khushalsagar@chromium.org

Change-Id: Id2fa0eade0fc12b81b5076a969e393410c41f7c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4570904
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Vladimir Levin <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1149779}
NOKEYCHECK=True
GitOrigin-RevId: 3c8e329bdb15e5b58e6d80cd9d77741a01b0ab80