[GridNG] Prevent zero-sized set track counts at non-zero indices
This CL fixes a a divide-by-zero in `ComputeTrackSizeRepeaterForRange`,
but the fix is nowhere near that code. The basic premise being violated
is that sets collections should always have the first set with a
`track_count` of 0, and no other indices should have a `track_count` of
zero. This is because the first set isn't actually a track
representation, it's for accumulated extra margin.

This premise gets violated when copying the set values from the parent
with opposite-direction children, which I added in
https://chromium-review.googlesource.com/c/chromium/src/+/4484336

The proper mitigation for this is to take the offset from the current
index, but the track count from the *next* index for opposite direction
subgrids. This will properly account for the initial entry with zero
tracks.

Bug: 1445565
Change-Id: I25586835034028f594895fa052d238a389383928
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4559196
Reviewed-by: Ethan Jimenez <ethavar@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1150006}
NOKEYCHECK=True
GitOrigin-RevId: 88b9eed6a1202c4d88985b57fc96b067c6c00d2b