Update VideoTrackAdapter frame rate reduction logic
VideoTrackAdapter drops frames in order to make sure that the output frame rate is close to the target frame rate despite an arbitrary
input frame rate.

However, too many frames were dropped occasionally. After digging into this I've found that the logic was very sensitive to jitter in the
input.

If the input timestamps are 10 fps with the following pattern:
0, 10 ms, 200 ms, 210 ms, 400 ms, 410 ms, 600 ms, 610 ms,...

and the max fps is set to 5 fps, VideoTrackAdapter produced something
like this:
0, 400 ms, 1000 ms, 1800 ms, 2610 ms, 3800 ms, 4800 ms, 5800 ms,...

which is roughly 1 fps. With the current CL, the output is as expected:
0, 200 ms, 400 ms, 600 ms, 800 ms, etc.

Bug: chromium:1448046
Change-Id: I71289a7a3986299c447e3d505cf07b52641e4aff
Low-Coverage-Reason: Removing unused field.
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4555847
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Commit-Queue: Johannes Kron <kron@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Henrik Boström <hbos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1150071}
NOKEYCHECK=True
GitOrigin-RevId: affcc0f6a82271ace23d37e29f62c9576bcfad70