Add a canvas_type param that can be used by Jinja templates.
This allows the template to know that canvas type the test is being
rendered for. For instance, templates can now do:
  {% if canvas_type == 'htmlcanvas' %}

For this change to work, the rendering of the code has to be moved
to a latter point, after the canvas_type is known and added to the
params.

Change-Id: I10345785b14bfe5c62bc848098a23c9f9806ee24
Bug: 1275750
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4567565
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Reviewed-by: Yi Xu <yiyix@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1150307}
NOKEYCHECK=True
GitOrigin-RevId: c3e777da9d8edda23d058e62a056a40b2b429c42