Migrate calls from SkSurface::flush -> GrDirectContext::flush
Presently, SkSurface::flush() defers to the GrDirectContext used
to create the surface to send work to the GPU. As part of an effort
to decouple SkSurface from the GPU code, we want to explicitly
call flush on the GrDirectContext and remove flushing from Surfaces.
http://review.skia.org/698237

In migrating these APIs, I tried to track where that context was
and use it directly and explicitly. In cases where that was not
obvious, I fell back to the existing behavior of "extract
GrRecordingContext, cast to GrDirectContext, check if it is non-null
(e.g. Raster or Graphite surfaces), and then call flush". In cases
where the default flush options were used, the helper
skgpu::ganesh::Flush() was used.

One peculiarity I found was in skia_output_device_dawn.cc where
the surfaces made in that file appear to be Graphite-backed
(and thus calling flush is a no-op), but when I removed that flush,
tests broke. So, I kept flushing in that spot.

Bug: skia:13983
Change-Id: Ifc68815b5da4b1a0d53099d1154440600e59175d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4566404
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Kevin Lubick <kjlubick@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1150871}
NOKEYCHECK=True
GitOrigin-RevId: 6abb9c6ab1f716ed3ff64794ed51fb85b7453c45