[@container] Always create ContainerQueryEvaluator on demand
Creating them eagerly for size containers and lazily for style
containers made the code confusing and hard to reason about.

Also there was a bug with incorrect evaluation of style queries for
size containers with unknown results for it size features. That is
fixed by making sure we do not create ContainerQueryEvaluators without a
MediaQueryEvaluator.

Bug: 1450054
Change-Id: I9af679badaff645da3d2b20a816821a7d1b266f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4567559
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151733}
NOKEYCHECK=True
GitOrigin-RevId: 159bc57740ee3779bf695433c590cbb6c9a003ab