[Auto-disable Accessibility] Remove aria-touchpassthrough logic
This CL removes the aria-touchpassthrough logic, which did not make it
into the aria spec.

The touchpassthrough concept is interesting, but the traction on it has
stalled, and implementation details are still being debated. The logic
added in Android created a regression, and since it is not being used
anywhere, we have decided to remove it rather than trying to make it
fit into the current framework when we don't even know the final spec
for the attribute, if there ever is one.

aria ticket: https://github.com/w3c/aria/issues/1215

AX-Relnotes: N/A
Bug: 1430202, b/265493191
Change-Id: I1d82ea384761db271abead5796171e7987ba9c76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4544402
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Bo Liu <boliu@chromium.org>
Reviewed-by: Mustaq Ahmed <mustaq@chromium.org>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1151986}
NOKEYCHECK=True
GitOrigin-RevId: a39a8f793f2072bf5c7022496f1cd03d25a03a17