[rebaseline-cl] Copy baselines for physical-under-virtual tests
Physical tests under virtual directories generate malformed optimization
graphs because different parts of the copier did not agree whether such
tests are virtual. To be consistent, ensure these tests are not virtual
for the purposes of optimization by checking for the existence of a test
file.

The optimizer is internally consistent, but add test coverage for it as
well.

Fixed: 1450725
Test: git cl issue 4577761
Test: ./blink_tool.py rebaseline-cl --no-trigger-jobs --patchset=3 -v
Test: New baselines under mac-mac12 and mac-mac13-arm64 that were not
Test: copied previously.
Change-Id: I1bf44458345276970107a9654bce82eb7bc11f6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4581826
Commit-Queue: Jonathan Lee <jonathanjlee@google.com>
Reviewed-by: Weizhong Xia <weizhong@google.com>
Cr-Commit-Position: refs/heads/main@{#1152253}
NOKEYCHECK=True
GitOrigin-RevId: 48c758d01efaa4576243f89ce53e22987a3ab951