WebNN: Implement the build method of split operation
This CL implements the MLGraphBuilder.split() method. The implementation
validates the input operand and options according to WebNN spec. It
implements the case where the splits parameter is an unsigned long or a
sequence of unsigned long. And it uses checked math to calculate the
output sizes.

This CL also adds unit tests that check the output shape calculation.

Bug: 1273291
Change-Id: I305402fb23977ceb844692b3144337964b86c14b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4420165
Reviewed-by: ningxin hu <ningxin.hu@intel.com>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Reviewed-by: Jiewei Qian <qjw@chromium.org>
Commit-Queue: Bin Miao <bin.miao@intel.com>
Cr-Commit-Position: refs/heads/main@{#1152309}
NOKEYCHECK=True
GitOrigin-RevId: b6ea2531b15272536b7f2185d52764402c92186b