[CSP] Fix CSP check for strict-dynamic with kSuppressReporting
The CSP checks take a parameter reporting_disposition which allows to
suppress reporting. The checks should return the same value
independently of that. Because of a mistake,
CSPDirectiveListAllowFromSource was behaving differently for a script
with url https://url.test in the case CSP: script-src 'strict-dynamic'
https://url.test depending on the reporting disposition.

This CL fixes that and prevents future mistakes by unifying the
checking logic. Unit tests are expanded to always check both variants.

Bug: 1299199
Change-Id: I048e028f74a520ec887232f06f4aa0933f140eb7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4575547
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Antonio Sartori <antoniosartori@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1152388}
NOKEYCHECK=True
GitOrigin-RevId: 9803fbb08f01dbfdb3e9e71e453edbb427fb3e4b