Implement "BeforeMatch" behavior in AnnotationAgent
A text fragment targeting text in a hidden but showable subtree will
cause the subtree to become visible. E.g. if the text is in a <details>,
has content-visibility: auto, or has the hidden=until-none attribute.

This behavior is currently implemented in TextFragmentAnchor. This CL
moves the behavior into AnnotationAgentImpl so it is more widely
available to other users of the Annotation API (e.g. user-notes, Chrome
Companion).

As part of this, we do a bit of cleanup and refactoring. The most
notable is that we no longer force an async break after text matching
unless the match is in a hidden subtree (so we can remove all the
double-rAFs from tests.

Motivation: this behavior is desired in Chrome Search Companion since
they use the kTextFinder annotations to perform a search of the document
to check if a text selector "works" in order to then offer up a text
directive link. This behavior is currently broken since the kTextFinder
"succeeds" as a match is found, however, the offered link does not work
since AnnotationAgent cannot open a hidden=until-none section.

This CL additionally makes kTextFinder side-effect free (doesn't
activate any locked DOM, doesn't create document markers) to make this
use case work.

Bug: 1379741, b/283857616
Change-Id: I272ae9b7431a4b34a65d1c3203b48f4bb835f613
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3930653
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1152500}
NOKEYCHECK=True
GitOrigin-RevId: ef86530afaba96641ef79bca11c397064cb0f780