Moving drawing_buffer to use TexImage instead of TexStorage
TexStorage is an extension in GLES2 and not guaranteed to exist.
WebGL1 is on the GLES2 context so we cannot assume that storage is
always available. It is also not safe to emulate texture storage in
angle because of complications related to storage support of BGRA
formats. [1]

[1]
https://source.chromium.org/chromium/chromium/src/+/main:gpu/command_buffer/service/feature_info.cc;l=919;drc=d04c248f2761ac4a96790fc89481eda5dad9a0ea

Bug: 1443160
Change-Id: I03126edc03aad6c9ff09fb614f138ec6040c6cfa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4547399
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Peter McNeeley <petermcneeley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1152543}
NOKEYCHECK=True
GitOrigin-RevId: b477d736ef9a1f4810809eeefc2df443f6356b9c