Add unit tests for FontSizeAdjust.
This change adds unit tests to verify the public APIs of FontSizeAdjust.
The tests examine the support of the two-value syntax of
font-size-adjust, which is newly introduced in CSS Font Module Level 5.
In addition, this change makes FontSizeAdjust::ToString truncate
trailing zeros if the return value is a floating number.

Bug: 1219875
Change-Id: I561a1cf110e5322cda77bd0035f083c08e271f07
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4582509
Commit-Queue: ChangSeok Oh <changseok.oh@bytedance.com>
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1152809}
NOKEYCHECK=True
GitOrigin-RevId: 998afed963affe067cd92f39cfb02ae358f5cf4f