Fix Spelling and Grammar overflow properly
The original fix for spelling and grammar text
decoration overflow did not invalidate overflow
when markers were added and deleted (the tests passed
because they added the markers before the first
lifecycle). This change fixes them properly by
invalidating overflow when the markers are added or
removed for any reason. It also simplifies the code
for target-text, and addresses potential problems
when markers are moved between nodes and when a node
does not have computed style.

Bug: 1433400,1450948
Change-Id: Ibbe88768d8fe2179ed40122f75e7f346cc33c4ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4583532
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1152938}
NOKEYCHECK=True
GitOrigin-RevId: 8bdd430b1e0b89add63fa8a8fd6a9a3c8d681208