Move the ServiceWorkerRouterRules typemap to shared_typemaps
To make the type maps also usable from the blink variant, this code
should be in shared_typemaps instead of cpp_typemaps.
We use this not only in the browser process, but also the renderer
process, and we need this type maps in both places.

Bug: 1371756
Change-Id: I0bc7e9bc3f0bc1f6b99fd8e9167655817b445dfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4583817
Reviewed-by: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1153075}
NOKEYCHECK=True
GitOrigin-RevId: 505f2e47bf96c52d2ed3f1f6683702785d59d029