CHECK (not DCHECK) avifImageSetViewRect succeeds
The image `view` is created as an empty image and passed to the
avifImageSetViewRect(view.get(), image, &rect) call. If
avifImageSetViewRect fails, then `view` is still an empty image and
it doesn't seem safe to use it.

The only reason avifImageSetViewRect can fail is invalid parameters.
Here the caller passes valid parameters, so this avifImageSetViewRect
should not fail.

Change-Id: Icc160905cbc17e622d4149fe690f257d7cec1803
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4582482
Commit-Queue: Wan-Teh Chang <wtc@google.com>
Reviewed-by: Peter Boström <pbos@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1153521}
NOKEYCHECK=True
GitOrigin-RevId: 9560acfa2c6ddd1b7b36bffc261dd49b56be7705