Fix under-invalidation when dynamically adding aria-describedby nodes
When an element with aria-describedby is added, we should call
AXRelationCache::UpdateReverseTextRelations to add the list
of ids referenced by aria-describedby, so that if elements receive
that id in the future, the description can be noted in the a11y tree.

This wasn't working unless there was already at least one node already
in the tree at the time the aria-describedby element is processed.
Now we call AXRelationCache::UpdateReverseTextRelations always,
which will fill in id_attr_to_text_relation_mapping_ with a key for
each attribute and an empty set value. This allows subsequent code
adding elements with those ids to notice that they should invalidate
the aria-describedby element.

In addition to the testcase in this CL, this fixes a source of
flakiness in CL 4553316.

Bug: 1342801

Change-Id: I3bcdf2659b22819ae37bf853d65f9464f2e0e2af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4592049
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1153910}
NOKEYCHECK=True
GitOrigin-RevId: b9d746aac9143d496ab677ea664352ec1891df36