media: Add "Copy video frame" support
This CL adds a context menu item "Copy video frame" to the <video>
element which copies the current video frame to the clipboard for
easier pasting and sharing.

Manually tested on Windows and Mac and the video frame is copied to
clipboard. On Windows, it also works when using hardware decoder
(D3D11VideoDecoder).

Also tested with protected content and nothing is copied to the
clipboard. In the next CL we'll disable this option for protected
content.

Also tested on Windows using MediaFoundationRenderer. In frame-server
mode it works as expected. In direct-composition mode, the copied
video frame is all black. Will discuss with owners and see what's the
best option for this case.

Bug: 1447318
Test: See above
Change-Id: I08359926f65cdf3084d6f7eb73ab77613ac8dddb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4550250
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: David Schinazi <dschinazi@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1153990}
NOKEYCHECK=True
GitOrigin-RevId: ecac3938ad3a8c0edd2d1a45c2f9c3c7a7a41c2b