[wptrunner] Support early exits
This change ports `run_web_tests.py --exit-after-n-*` flags to
`run_wpt_tests.py` to fail quickly if the build is unlikely to pass. The
results processor listens for regressions and sends a termination signal
to the main process if a threshold is exceeded.

A future improvement is to default to a proportional threshold, as
`run_web_tests.py` does, instead of disabling this feature by default.

This change depends on web-platform-tests/wpt#40086 to work well.

Bug: 1382175
Cq-Include-Trybots: luci.chromium.try:linux-wpt-identity-fyi-rel
Change-Id: I626458af74572b1439596e5f4eaf82523bd3d7b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4548601
Reviewed-by: Weizhong Xia <weizhong@google.com>
Commit-Queue: Jonathan Lee <jonathanjlee@google.com>
Cr-Commit-Position: refs/heads/main@{#1154474}
NOKEYCHECK=True
GitOrigin-RevId: ad0a3ae7c61912897b6909a1185f365eec9c8a17