Add SVGElement::SynchronizeAllSVGAttributes() for the "sync all" case
The various implementations of `SynchronizeSVGAttribute()` are basically
only handling the "sync all" case, and the base implementation in
`SVGElement::SynchronizeSVGAttribute()` handles the "single" case by
calling `PropertyFromAttribute()` and then synchronizing the attribute
object that's returned. De-virtualize the "single" case and let a new
`SynchronizeAllSVGAttributes()` handle the "sync all" case. This is
essentially achieved by renaming the existing variant and removing the
checks on the `name` parameter.

Rename the `SynchronizeAllSVGAttributes()` helper to
`SynchronizeListOfSVGAttributes()` to help differentiate between the
different variants. Use the helper in a few more instances.

Change-Id: I739a9005d4484d81c9b6b53c1c7365e960a9442c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4598007
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/main@{#1154857}
NOKEYCHECK=True
GitOrigin-RevId: f4334f36856a4500fd1cf50cb2d009e979e520fb