[LCPP] Notify LCPP on LCP update
This CL makes LCPCriticalPathPredictor to be notified when LCP element
is updated and prepares a locator string indicating the relative position
of the LCP element.

We plan to soon follow up with a CL to plumb the information to the browser
process.

Bug: 1419756
Change-Id: Iaba42641119f0d4b1fde54d57dd7edb85c91f923
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4600254
Auto-Submit: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1154878}
NOKEYCHECK=True
GitOrigin-RevId: 1bb4539036a27dea1a5ed379e690d08289df244d