Reland "Reland "Pass soft navigation start time and others to browser""
This is a reland of commit 59194d3467e54b4c5a50dbdab19cda4c4f19eeb9

The previous submission is reverted because of the added browser
test is flaky in the CI bot Linux Test. The test waiter which waits
for soft navigation detections times out in the failed cases.

This reland fixes the browser test by moving the waiter->Wait()
before navigating away to a blank page. This is the correct order.
The test waiter is an observer on web contents. It should wait before
the web contents being destroyed. It should be some timing issue that
the previously the test could also work on most bots.

This fix has passed CI bots and swarming bots.

Bug: 1370937
Change-Id: I3fd3b2a050a0ffcedbffa0af413f51d70e6326b2
Include-Ci-Only-Tests: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4594437
Commit-Queue: Hao Liu <haoliuk@chromium.org>
Reviewed-by: Sun Yueru <yrsun@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1155054}
NOKEYCHECK=True
GitOrigin-RevId: e358011944c3d4539d1911dd267ec23a6e3fd282